+++ wildcard/lockdown/testing.py
@@ -9,12 +9,19 @@ from plone.app.testing.layers import FunctionalTesting
 from plone.app.testing.layers import IntegrationTesting
 from zope.configuration import xmlconfig
 from plone.testing import z2
+from plone import api
+
+
+IS_PLONE_5 = api.env.plone_version().startswith('5')
 
 
 class Lockdown(PloneSandboxLayer):
     defaultBases = (PLONE_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
+        if IS_PLONE_5:
+            import plone.app.contenttypes
+            self.loadZCML(package=plone.app.contenttypes)
         # load ZCML
         import wildcard.lockdown
         xmlconfig.file('configure.zcml', wildcard.lockdown,
@@ -22,6 +29,9 @@ class Lockdown(PloneSandboxLayer):
         z2.installProduct(app, 'wildcard.lockdown')
 
     def setUpPloneSite(self, portal):
+        if IS_PLONE_5:
+            applyProfile(portal, 'plone.app.contenttypes:default')
+
         # install into the Plone site
         applyProfile(portal, 'wildcard.lockdown:default')
         setRoles(portal, TEST_USER_ID, ('Member', 'Manager'))